Fix extra spacing between some skills
Because the skill ladders are `inline-block` to maintain some space at the start/end of columns the margins of skills in adjacent ladders do not overlap, instead double the margin between adjacent skills in the same ladder so they're all consistent.  I tried removing the bottom margin of skills instead, but this then leaves no space at the bottom of a column.