Rename Guilds to Lists
We didn't have a better name for these when we made the first version of
the site, but "lists" better describes what they are (Occupational Skill
Lists).

All references to "guild" used in this way have been changed; but
genuine uses of "Guild" have been kept.

I've also reformatted the TOML lists in every skill.